Base Project
Just created a simple temperature converter. Main advantage with this is that just practiced events and how to take input and convert data from string to number.
By default data that we get through html input fields are a string, so we need to convert it into number to work  with it.